Revert "Thinh improve (#27)"
This reverts commit edc5487248f2d5fe1152eab277a4e4b219495ae7.
