--- Assets/Instruction.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6b625254a7878411fbc324d1697644cf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
