--- Assets/Instruction/PauseGame_Menu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a46ac0893e78348f68435cce4be4c953
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
