--- Assets/Materials/ColorFront.asset
@@ -1,19 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 54d21f6ece3b46479f0c328f8c6007e0, type: 3}
-  m_Name: ColorFront
-  m_EditorClassIdentifier: 
-  colorMode: 3
-  topLeft: {r: 0.6, g: 0.2627451, b: 0.34901962, a: 1}
-  topRight: {r: 0.6, g: 0.2627451, b: 0.34901962, a: 1}
-  bottomLeft: {r: 0.5294118, g: 0.25490198, b: 0.627451, a: 1}
-  bottomRight: {r: 0.7176471, g: 0.6784314, b: 0.3019608, a: 1}
