--- Assets/Materials/ColorFront.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: defb3b5112bad40bf85db3a9e861889a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
