+++ Assets/Scripts/AppEvents/Bear/BearFootStepEmitter.cs
@@ -7,7 +7,7 @@ public class BearFootStepEmitter : MonoBehaviour
 {
     public void Start()
     {
-        InvokeRepeating("ExecuteFootstep", 2.0f, 13.0f); // Executing every 13s
+        InvokeRepeating("ExecuteFootstep", 2.0f, 10.0f); // Executing every 10s
     }
 
     private void ExecuteFootstep() {
