--- Assets/Scripts/DynamicSliderAdjust.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5440634b7370e41b9af68cd5035d9cb8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
