--- Assets/Scripts/DynamicSliderAdjust/InGameCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9ebac365d947841a5b4f2a486f85f1ff
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
