--- Assets/Scripts/DynamicSliderAdjust/InGameCharVol.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8bfb999755840483d9afc79664d53ec5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
