--- Assets/Scripts/DynamicSliderAdjust/InGameVolume.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 616cfb82e1a9b4821a5b0d44c36c9070
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
