+++ Assets/Scripts/DynamicSoundAdjust.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 4a9618cbe39c04c1491fb855838e62ca
-PrefabImporter:
+guid: b6f3519109bf5464e8adb5b63bc6f6d3
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
