--- Assets/Scripts/EventSystem_Menu/GameInstruction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07155dd74f2504ae39d2a91fccbed86b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
